feat: add types to resolvers
