+++ apps/server/codegen.yml
@@ -1,0 +1,8 @@
+schema: "./src/schema.graphql"
+generates:
+  ./src/__generated__/resolvers-types.ts:
+    plugins:
+      - "typescript"
+      - "typescript-resolvers"
+    config:
+      useIndexSignature: true
