+++ apps/server/src/resolvers/mutations.ts
@@ -1,0 +1,57 @@
+import {
+  MutationResolvers,
+  ReportInput,
+  UserInput,
+} from '../__generated__/resolvers-types'
+import Report from '../lib/mongoose/models/Report'
+import User from '../lib/mongoose/models/User'
+
+const mutations: MutationResolvers = {
+  Mutation: {
+    createUser: async ({ input }: { input: UserInput }) => {
+      try {
+        const newUser = new User(input)
+        const savedUser = await newUser.save()
+        return savedUser
+      } catch (error) {
+        throw new Error(
+          'Error creating a new user and saving it to the database'
+        )
+      }
+    },
+    updateUser: async ({ input }: { input: UserInput }) => {
+      try {
+        const updatedUser = await User.findOneAndUpdate(
+          { email: input.email },
+          input,
+          { new: true }
+        )
+        return updatedUser
+      } catch (error) {
+        throw new Error('Error updating a user in the database')
+      }
+    },
+
+    updateReport: async ({
+      targetId,
+      cycleId,
+      input,
+    }: {
+      targetId: String
+      cycleId: String
+      input: ReportInput
+    }) => {
+      try {
+        const filter = { '_id.target': targetId, '_id.cycle': cycleId }
+        const updatedReport = await Report.findOneAndUpdate(filter, input, {
+          new: true,
+        })
+        return updatedReport
+      } catch (error) {
+        throw new Error('Error updating a report in the database')
+      }
+    },
+  },
+}
+
+export default mutations
